pyproject.toml
[build-system]
requires = ["setuptools>=60.2.0", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "financial_tracker"
version = "0.1.0"
description = ""
readme = "README.md"

dependencies = [

"pandas",
"psycopg2-binary",
"sqlalchemy"

]



